feat(docparser): support jsonapi tags
The specification format is available at http://jsonapi.org/format/
Support `jsonapi` tags during the generation phase.

Support status of `top-level` members

- `data` is supported,
- `includes` is supported (skip it with `--without-jsonapi-includes`),
- `errors` objects can be handle by `json` tags
(with an `ErrorObject` and an `ErrorsPayload` components).
- `meta`, `jsonapi`, `links` are not supported.

For a given `Foo` object, the use of `jsonapi` tags will generated
all components needed to represent the `jsonapi` format of this object.
- `Foo` as a resource object
- `FooAttributes`
- `FooCollection`: as an array of resource objects
- `FooData`
- `FooIdentifier`: as a resource identifier object
- `FooIdentifierCollection`: as an array of resource identifier objects
- `FooIdentifierData`
- `FooRelationships`

Others contents needed for this support
- `make test` will run openapi-parser against various datasets,
- Support of `string` attribute in a `json` tag (`dataset/jsontags`),
- Fix a typo in `yaml` tag of the `server` struct,
- Parse `json`, `jsonapi`, and `validate` tags separately,
- Field `Items` from `Schema` now support
a single string value OR a array of `Schema`,
- All warnings during file parsing became fatal
to the openapi file generation.